Refs #2287. Merges mythtv-vid to trunk.
There are a number of changes, probably the most user visible ones are PiP related, but there is also groundwork laid for OpenGL video rendering on OSX and MS Windows, for aspect ratio & zoom on OSX, and a number of operations are more responsive.

Note: Even though this changeset has received a fair bit of testing, it is a 1.1MB patch and there are bound to be things we overlooked. Please report any compilation problems to the mythtv-dev mailing list and open tickets with a repro description, theme used and backtrace for any segfaults or deadlocks.

git-svn-id: http://svn.mythtv.org/svn/trunk@19417 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/remoteencoder.h
@@ -32,7 +32,7 @@ class MPUBLIC RemoteEncoder
     long long GetFreeDiskSpace();
     long long GetMaxBitrate();
     long long GetKeyframePosition(long long desired);
-    void FillPositionMap(int start, int end,
+    void FillPositionMap(long long start, long long end,
                          QMap<long long, long long> &positionMap);
     void StopPlaying(void);
     void SpawnLiveTV(QString chainid, bool pip, QString startchan);
